Refactor form component
