+++ src/components/index.js
@@ -1,2 +1,1 @@
 export { default as MapGL } from "./map-gl/MapGL";
-export { default as Form } from "./form/Form";
