+++ src/views/form/constants.js
@@ -1,0 +1,8 @@
+const { freeze } = Object;
+
+export const FORM_INFORMATION = 'To display locations correctly, drag and drop below fields to their representative column headers.';
+
+export const BUTTON = freeze({
+  ACCEPT: 'Accept',
+  CANCEL: 'Cancel',
+});
