+++ src/views/index.js
@@ -1,1 +1,2 @@
 export { default as UploadFileView } from "./upload-file/UploadFileView";
+export { default as Form } from "./form/Form";
